♻️ Refactor: 명언 피드 컴포넌트 이름 정리
