+++ src/components/QuoteFeed/QuoteFeed.styles.ts
